requirements.txt
numpy
scipy
scikit-learn
pillow
matplotlib